+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* timeline colors can be configured in `rtw_config.json`
 * add `day` subcommand (display timeline for the current day)
 * add `week` subcommand (display timeline for the current week)